Generate local-ancestry-aware GWAS results

This function uses a glmmkin class object from the null GLMM to perform score tests for local-ancestry-aware association with genotypes in a GDS file .gds file.

get.la.aware.score(
  null.obj,
  geno.file,
  outfile,
  LAC.file,
  n_PC_used = NULL,
  id.type = c("character", "integer", "numeric"),
  MAF.range = c(1e-07, 0.5),
  miss.cutoff = 1,
  missing.method = "impute2mean",
  is.dosage = F,
  n.batch = 100
)

Arguments

null.obj

a class glmmkin or class glmmkin.multi object, returned by fitting the null GLMMusing glmmkin.

geno.file

the full name of a GDS file (including the suffix .gds), or an object of class SeqVarGDSClass.

outfile

the output file name.

LAC.file

the full name of a binary LAC file.

n_PC_used

the dimension of PCs that will be included in analysis. Default is set to the number of dimension of PCs in the LAC.file.

id.type

the class of sample IDs. Possible values are "character", "integer", and "numeric".

MAF.range

a numeric vector of length 2 defining the minimum and maximum minor allele frequencies of variants that should be included in the analysis (default = c(1e-7, 0.5)).

miss.cutoff

the maximum missing rate allowed for a variant to be included (default = 1, including all variants).

missing.method

method of handling missing genotypes. Either "impute2mean" or "omit" (default = "impute2mean").

is.dosage

a logical switch for whether imputed dosage should be used from a GDS infile (default = FALSE).

n.batch

an integer for how many SNPs should be tested in a batch (default = 100). The computational time can increase dramatically if this value is either small or large. The optimal value for best performance depends on the user’s system.

Value

@format a dataframe with following components:

SNP

SNP name, as supplied in snps.

CHR

Chromosome, copied from .gds file.

POS

physical position in base pairs, copied from .gds file.

REF

reference allele, copied from .gds file.

ALT

alternate allele, copied from .gds file.

MISSRATE

number of individuals with non-missing genotypes for each SNP.

AF

ALT allele frequency for each SNP.

N

total sample size.

SCORE

the summary score of the effect allele.

VAR

the variance of the summary score.

PVAL

local-ancestry-aware GWAS p-values.