generate local-ancestry-aware phenotype distribution plots

This function generates utilizes ggplot2 to generate phenotype or covariate distribution in the context of local ancestry at a specfic genetic position and global PCs as reference. The proportion of phenotype subtypes will be distributed in pie chart ratios, and the size of pie chart indicates the number of samples sharing that specific local ancestry combination. Only works for categorical variables.

# S3 method for class 'features'
plot(
  input.file,
  PC1,
  PC2,
  anc,
  PC1_la,
  PC2_la,
  pheno,
  pheno.name,
  phenotype.name,
  y.scale = 1e+06,
  if.jitter = T,
  y.jitter = 0.01,
  if.round = T,
  round.digit = 2,
  if.filter = F,
  n.filter = 5,
  r.x.axis = -0.025,
  r.y.axis = 0.025,
  r.n.level = 4
)

Arguments

input.file

a dataframe that includes information of global PCs, LACs at genetic positions and phenotype or covariate of interest.

PC1, PC2

a pair of strings indicating the columns contributed to the x-, and y-coordinate in the background, which are usually supposed to be from global PCs, e.g. PC1 and PC2. Please do not include any single quote nor double quote surrounding the string.

anc

a vector indicating reference ancestry populations.

PC1_la, PC2_la

a pair of string indicating the columns corresponding to x-, and y-coordinate in LACs, e.g. PC1_la and PC2_la, on which the combination of local ancestry and phenotype subtypes will be counted. Please do not include any single quote nor double quote surrounding the string.

pheno

a string indicating of the column of phenotype or covariate of interest.

pheno.name

a string as the label of the phenotype or covariate.

phenotype.name

a string vector as labels of phenotype subtypes.

y.scale

an integer or a positive value, usually extreme large, adjusting the radius of pie charts to match the coordinates of PCs (default = 1e6).

if.jitter

a logical switch indicating whether to slightly jitter the y-coordinates of pie charts to reduce overlapping between local ancestry patterns and pie charts (default = TRUE).

y.jitter

a positive number to adjust the y-coordinates of pie charts (default = 0.01). Only effective when if.jitter = TRUE.

if.round

a logical switch indicating whether to round LACs to reduce the impact of deviation in the LACs (default = TRUE).

round.digit

an integer indicating the number decimal places to be used (default = 2). Only effective when if.round = TRUE.

if.filter

a logical switch indicating whether to filter the number of local ancestry patterns to reduce the impact of deviation in the LACs (default = TRUE).

n.filter

an integer indicating the number of threshold for local ancestry patterns preserving for further analysis (default = 5). Only effective when if.filter = TRUE.

r.x.axis, r.y.axis

a pair of numerics indicating the coordinates of legend of pie chart sizes (default = c(-0.025,0.025)).

r.n.level

an integer indicating the layers of circles in the legend of pie chart sizes (default = 4).

Value

a ggplot2-based plots that illustrates the relationship between global PCs, LACs and phenotype distribution per local ancestry pattern accordingly.